+++ openhands/integrations/github/service/__init__.py
@@ -4,7 +4,6 @@ from openhands.integrations.github.service.branches_prs import (
     GitHubBranchesMixin,
     GitHubPRsMixin,
 )
-from openhands.integrations.github.service.core import GitHubCoreMixin
 from openhands.integrations.github.service.graphql import GitHubGraphQLMixin
 from openhands.integrations.github.service.microagents import GitHubMicroagentsMixin
 from openhands.integrations.github.service.repos import GitHubReposMixin
@@ -16,7 +15,6 @@ from openhands.integrations.service_types import (
 
 
 class GitHubService(
-    GitHubCoreMixin,
     GitHubReposMixin,
     GitHubBranchesMixin,
     GitHubPRsMixin,
@@ -37,12 +35,20 @@ class GitHubService(
         external_token_manager: bool = False,
         base_domain: str | None = None,
     ) -> None:
-        GitHubCoreMixin.__init__(
-            self,
-            user_id=user_id,
-            external_auth_id=external_auth_id,
-            external_auth_token=external_auth_token,
-            token=token,
-            external_token_manager=external_token_manager,
-            base_domain=base_domain,
-        )
+        self.user_id = user_id
+        self.external_token_manager = external_token_manager
+
+        if token:
+            self.token = token
+
+        if base_domain and base_domain != 'github.com':
+            self.BASE_URL = f'https://{base_domain}/api/v3'
+
+        self.external_auth_id = external_auth_id
+        self.external_auth_token = external_auth_token
+
+    @property
+    def provider(self) -> str:
+        from openhands.integrations.service_types import ProviderType
+
+        return ProviderType.GITHUB.value
