--- openhands/integrations/github/service/core.py
@@ -1,36 +1,0 @@
-from pydantic import SecretStr
-
-
-class GitHubCoreMixin:
-    """Core mixin for GitHub service shared state and configuration."""
-
-    BASE_URL = 'https://api.github.com'
-    token: SecretStr = SecretStr('')
-    refresh = False
-
-    def __init__(
-        self,
-        user_id: str | None = None,
-        external_auth_id: str | None = None,
-        external_auth_token: SecretStr | None = None,
-        token: SecretStr | None = None,
-        external_token_manager: bool = False,
-        base_domain: str | None = None,
-    ) -> None:
-        self.user_id = user_id
-        self.external_token_manager = external_token_manager
-
-        if token:
-            self.token = token
-
-        if base_domain and base_domain != 'github.com':
-            self.BASE_URL = f'https://{base_domain}/api/v3'
-
-        self.external_auth_id = external_auth_id
-        self.external_auth_token = external_auth_token
-
-    @property
-    def provider(self) -> str:
-        from openhands.integrations.service_types import ProviderType
-
-        return ProviderType.GITHUB.value
